+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- \[mpm\] Keep the operation matrix on the `readme.md` in sync with current code by inspecting implementation.
+
 ## {gh}`5.4.0 (2022-06-28) <compare/v5.3.0...v5.4.0>`
 
 - \[mpm\] Allow global `upgrade` of a subset of packages from the command line if no ambiguity is identified.